--- src/components/musicalCube/Toolbar.css
@@ -1,7 +1,0 @@
-.cube-side-toolbar { 
-    margin-top: -35px;
-
-    ion-icon { 
-        color : var(--yellow-800);
-    }
-}
