+++ src/components/musicalCube/toolbar/Toolbar.css
@@ -1,0 +1,7 @@
+.cube-side-toolbar { 
+    margin-top: -35px;
+
+    ion-icon { 
+        color : var(--yellow-800);
+    }
+}
