+++ src/pages/realms/musicalCubes/sounds/MusicalCubeSounds.ts
@@ -1,6 +1,9 @@
-import pianoCm from "./84_keys_acoustic_strictly_piano_chords_Cmin.wav"
+import guitar_B from './guitar/AHA_SV_60_guitar_stack_loop_closer_now_B.wav';
+import melody_B from './melody/AHA_SV_60_melody_loop_closer_now_no_keys_no_bass_B.wav';
+import vocals_Am from './vocals/AHA_SV_60_vocal_loop_in_the_city_6-8_Am.wav';
 
 export enum MusicalCubeSounds {
-    PianoCmin = pianoCm,
-    PianoCopy = pianoCm,
+    guitarB = guitar_B,
+    melodyB = melody_B,
+    vocalsAm = vocals_Am
   }
